It now communicates with the arduino.
It now perfectly communicates with the arduino. Anything that you draw
on the program will show up on the display. I had to modify the saved
files for 4 bit colour as opposed to 8 bit colour. I also had to do this
for the colour selector and displaying the pixels. It updates once per
second. Now I need to add a fill tool and maybe an image converter.